requirements.txt
Pinyin2Hanzi==0.1.1
pypinyin==0.49.0
pytorch_crf==0.7.2
streamlit==1.20.0
torch==1.13.1
transformers==4.29.2
websocket_client==1.4.2
